+++ tests/test_datastructures.py
@@ -300,10 +300,6 @@ class _ImmutableDictTests:
             a |= {"y": 2}
 
 
-class TestImmutableTypeConversionDict(_ImmutableDictTests):
-    storage_class = ds.ImmutableTypeConversionDict
-
-
 class TestMultiDict(_MutableMultiDictTests):
     storage_class = ds.MultiDict
 
@@ -372,25 +368,6 @@ class TestMultiDict(_MutableMultiDictTests):
 
         with pytest.raises(KeyError):
             md["empty"]
-
-
-class TestTypeConversionDict:
-    storage_class = ds.TypeConversionDict
-
-    def test_value_conversion(self):
-        d = self.storage_class(foo="1")
-        assert d.get("foo", type=int) == 1
-
-    def test_return_default_when_conversion_is_not_possible(self):
-        d = self.storage_class(foo="bar", baz=None)
-        assert d.get("foo", default=-1, type=int) == -1
-        assert d.get("baz", default=-1, type=int) == -1
-
-    def test_propagate_exceptions_in_conversion(self):
-        d = self.storage_class(foo="bar")
-        switch = {"a": 1}
-        with pytest.raises(KeyError):
-            d.get("foo", type=lambda x: switch[x])
 
 
 class TestCombinedMultiDict: